Add basic type annotations to the QML/JS grammar
Inspired by TypeScript syntax, allow optional type annotations in the
style of ": <name of type>" in various places such as after function
declarations, after variable declarations and in function parameters.

There are some places where annotations are not permitted yet in the
grammar, for example for arrow function parameters. This will require
additional work.

These annotations are not used at the moment, so by default we produce
an error message when encountering them in the AST.

Change-Id: I12d20b4e2ce92e01108132975a06cfd13ba4a254